Create TextReader.sh
this is a shell script to go through all the statistics.txt files created by DSI_StudioCommand.py and extract the stats and put them up in the terminal for you to view.  There are more elegant ways to do this, but I am lazy and I spit most of it out and then format it in LaTeX or R.  

Again, it starts with the tract "dictionary" and then loads the corresponding file, pulls out the second column of the report and spits out the data.  It is built for the data found in the IXI database in the nitrc repository, hence the directory sub heading.